+++ tests/test_trash_launcher_icon.cpp
@@ -48,7 +48,7 @@ TEST_F(TestTrashLauncherIcon, Position)
 
 TEST_F(TestTrashLauncherIcon, Activate)
 {
-  unsigned long long time = g_random_int();
+  uint64_t time = g_random_int();
   EXPECT_CALL(*fm_, OpenTrash(time));
   icon.Activate(ActionArg(ActionArg::Source::LAUNCHER, 0, time));
 }
